functions/legacy/requirements.txt
firebase-functions
firebase-admin
pandas
finance-datareader
lxml
requests
bs4
schedule
python-dotenv
